requirements.txt
flask>1.1
flask_sqlalchemy
pymysql
pyjwt
flask_caching
